# docs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /api/admin/users/{userId}/role:
    put:
      tags:
        - admin-user-controller
      operationId: changeUserRole
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: role
          in: query
          required: true
          schema:
            type: string
            enum:
              - USER
              - ADMIN
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
...
  /api/admin/users/{userId}/unblock:
    post:
      tags:
        - admin-user-controller
      operationId: unblockUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
  /api/admin/users/{userId}/block:
    post:
      tags:
        - admin-user-controller
      operationId: blockUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
...
  /api/admin/users/{userId}:
    get:
      tags:
        - admin-user-controller
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
...
components:
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        cardHolder:
          type: string
        role:
          type: string
          enum:
            - USER
            - ADMIN
...